py/requirements.txt
mariadb
python-dotenv
